Refactor to minimise hubconf dependencies
