+++ hi-ml-multimodal/src/health_multimodal/image/model/__init__.py
@@ -5,9 +5,12 @@
 
 from .model import ImageModel
 from .model import ResnetType
-
+from .model import get_biovil_resnet
+from .model import CXR_BERT_COMMIT_TAG
 
 __all__ = [
     "ImageModel",
     "ResnetType",
+    "get_biovil_resnet",
+    "CXR_BERT_COMMIT_TAG",
 ]
